+++ sonar-delphi-plugin/src/main/java/au/com/integradev/delphi/DelphiPlugin.java
@@ -22,6 +22,7 @@
  */
 package au.com.integradev.delphi;
 
+import au.com.integradev.delphi.check.DefaultMetadataResourcePath;
 import au.com.integradev.delphi.core.Delphi;
 import au.com.integradev.delphi.coverage.DelphiCoverageParserFactory;
 import au.com.integradev.delphi.enviroment.DefaultEnvironmentVariableProvider;
@@ -51,7 +52,7 @@ public class DelphiPlugin implements Plugin {
     builder.add(
         // Core
         Delphi.class,
-        DelphiSonarWayProfile.class,
+        DelphiRulesDefinition.class,
         // Sensors
         DelphiSensor.class,
         // Executors
@@ -61,10 +62,15 @@ public class DelphiPlugin implements Plugin {
         // Core helpers
         DelphiProjectHelper.class,
         // Environment
-        DefaultEnvironmentVariableProvider.class);
+        DefaultEnvironmentVariableProvider.class,
+        // Paths
+        DefaultMetadataResourcePath.class,
+        DelphiSonarWayResourcePath.class);
 
     if (context.getRuntime().getProduct() == SonarProduct.SONARQUBE) {
       builder.add(
+          // Core
+          DelphiSonarWayProfile.class,
           // Sensors
           DelphiCoverageSensor.class,
           DelphiNUnitSensor.class,