+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Rename all rule keys. (For example, `ClassNameRule` is now `ClassName`)
 - Rename plugin key from `delphi` to `communitydelphi`.
 - Rename repository key from `delph` to `community-delphi`.
+- Rename repository from `Delphi` to `Community Delphi`.
 
 ### Removed